+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/repository/UserRepository.java
@@ -13,6 +13,6 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     @Query( "select u from User u join Role r on u.role.roleId = r.roleId where u.userName like %:name% and r.roleName ='LECTURER' ")
     List<User> findUserByUserName(String name);
 
-    User findByUserId(int id);
-
+    User findByUserId(Long id);
+//    List<User> findUserByEmptySlots_Empty(Long slotId);
 }
